+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.2.6 (TBA)
+
+* Added `Assent.HTTPAdapter.Finch`
+* Deprecated `Assent.HTTPAdapter.Mint`
+
 ## v0.2.5 (2023-08-21)
 
 * `Assent.Strategy.Spotify` added